refactor: cleaning up the utility classes + docs
